+++ src/io/trr_io.rs
@@ -9,7 +9,7 @@ use std::path::Path;
 
 use crate::errors::{ReadTrajError, TrajError, WriteTrajError};
 use crate::io::traj_io::{
-    FrameData, TrajGroupWrite, TrajRangeRead, TrajRead, TrajReader, TrajWrite,
+    FrameData, TrajGroupWrite, TrajRangeRead, TrajStepRead, TrajRead, TrajReader, TrajWrite,
 };
 use crate::io::xdrfile::{self, CXdrFile, OpenMode, XdrFile};
 use crate::iterators::AtomIterator;
@@ -121,7 +121,7 @@ impl<'a> TrajRead<'a> for TrrReader<'a> {
     fn new(
         system: &'a mut System,
         filename: impl AsRef<Path>,
-    ) -> Result<TrajReader<'a, TrrReader>, ReadTrajError> {
+    ) -> Result<TrrReader, ReadTrajError> {
         let n_atoms = system.get_n_atoms();
 
         // sanity check the number of atoms
@@ -148,7 +148,7 @@ impl<'a> TrajRead<'a> for TrrReader<'a> {
             phantom: PhantomData,
         };
 
-        Ok(TrajReader::wrap_traj(trr_reader))
+        Ok(trr_reader)
     }
 
     fn get_system(&mut self) -> *mut System {
@@ -167,6 +167,22 @@ impl<'a> TrajRangeRead<'a> for TrrReader<'a> {
                 Err(ReadTrajError::StartNotFound(start_time.to_string()))
             } else {
                 Ok(())
+            }
+        }
+    }
+}
+
+impl<'a> TrajStepRead<'a> for TrrReader<'a> {
+    fn skip_frame(&mut self) -> Result<bool, ReadTrajError> {
+        unsafe {
+            match xdrfile::trr_skip_frame(self.get_file_handle()) {
+                0 => Ok(true),
+                1 => Err(ReadTrajError::SkipFailed()),
+                2 => Ok(false),
+                number => panic!(
+                    "Groan error. `trr_skip_frame` returned '{}' which is unsupported.",
+                    number
+                ),
             }
         }
     }
@@ -204,8 +220,9 @@ impl System {
     ///
     /// You can also iterate over just a part of the trajectory.
     /// Here, only frames in the time range 10-100 ns will be read.
-    /// The `with_range` method is very efficient as the particle properties
-    /// from the trr frames outside of the range will not be read.
+    /// The `with_range` method is very efficient for the trr files
+    /// as the particle properties from the trr frames outside of 
+    /// the range will not be read.
     /// ```no_run
     /// use groan_rs::prelude::*;
     /// use groan_rs::errors::ReadTrajError;
@@ -216,6 +233,29 @@ impl System {
     ///     for raw_frame in system
     ///         .trr_iter("trajectory.trr")?
     ///         .with_range(10_000.0, 100_000.0)?
+    ///     {
+    ///         let frame = raw_frame?;
+    ///         println!("{:?}", frame.group_get_center("all"));
+    ///     }
+    ///
+    ///     Ok(())
+    /// }
+    /// ```
+    /// 
+    /// Furthermore, you can efficiently skip over some frames of the trajectory.
+    /// Here, only every 10th frame of the trajectory will be read.
+    /// The `with_step` method is very efficient for the trr files as the 
+    /// particle properties from the skipped over frames will not be read.
+    /// ```no_run
+    /// use groan_rs::prelude::*;
+    /// use groan_rs::errors::ReadTrajError;
+    ///
+    /// fn example_fn() -> Result<(), ReadTrajError> {
+    ///     let mut system = System::from_file("system.gro").unwrap();
+    ///
+    ///     for raw_frame in system
+    ///         .trr_iter("trajectory.trr")?
+    ///         .with_step(10)?
     ///     {
     ///         let frame = raw_frame?;
     ///         println!("{:?}", frame.group_get_center("all"));
@@ -237,7 +277,7 @@ impl System {
         &mut self,
         filename: impl AsRef<Path>,
     ) -> Result<TrajReader<TrrReader>, ReadTrajError> {
-        TrrReader::new(self, filename)
+        Ok(TrajReader::wrap_traj(TrrReader::new(self, filename)?))
     }
 }
 
@@ -540,6 +580,27 @@ mod tests {
     use float_cmp::assert_approx_eq;
     use std::fs::File;
     use tempfile::NamedTempFile;
+    use crate::structures::atom::Atom;
+
+    fn compare_atoms(atom1: &Atom, atom2: &Atom) {
+        assert_eq!(atom1.get_residue_number(), atom2.get_residue_number());
+        assert_eq!(atom1.get_residue_name(), atom2.get_residue_name());
+        assert_eq!(atom1.get_atom_number(), atom2.get_atom_number());
+        assert_eq!(atom1.get_atom_name(), atom2.get_atom_name());
+        assert_eq!(atom1.get_chain(), atom2.get_chain());
+
+        assert_approx_eq!(f32, atom1.get_position().x, atom2.get_position().x);
+        assert_approx_eq!(f32, atom1.get_position().y, atom2.get_position().y);
+        assert_approx_eq!(f32, atom1.get_position().z, atom2.get_position().z);
+
+        assert_approx_eq!(f32, atom1.get_velocity().x, atom2.get_velocity().x);
+        assert_approx_eq!(f32, atom1.get_velocity().y, atom2.get_velocity().y);
+        assert_approx_eq!(f32, atom1.get_velocity().z, atom2.get_velocity().z);
+
+        assert_approx_eq!(f32, atom1.get_force().x, atom2.get_force().x);
+        assert_approx_eq!(f32, atom1.get_force().y, atom2.get_force().y);
+        assert_approx_eq!(f32, atom1.get_force().z, atom2.get_force().z);
+    }
 
     #[test]
     fn read_trr() {
@@ -833,23 +894,7 @@ mod tests {
             let frame2 = raw_frame2.unwrap();
 
             for (atom1, atom2) in frame1.atoms_iter().zip(frame2.atoms_iter()) {
-                assert_eq!(atom1.get_residue_number(), atom2.get_residue_number());
-                assert_eq!(atom1.get_residue_name(), atom2.get_residue_name());
-                assert_eq!(atom1.get_atom_number(), atom2.get_atom_number());
-                assert_eq!(atom1.get_atom_name(), atom2.get_atom_name());
-                assert_eq!(atom1.get_chain(), atom2.get_chain());
-
-                assert_approx_eq!(f32, atom1.get_position().x, atom2.get_position().x);
-                assert_approx_eq!(f32, atom1.get_position().y, atom2.get_position().y);
-                assert_approx_eq!(f32, atom1.get_position().z, atom2.get_position().z);
-
-                assert_approx_eq!(f32, atom1.get_velocity().x, atom2.get_velocity().x);
-                assert_approx_eq!(f32, atom1.get_velocity().y, atom2.get_velocity().y);
-                assert_approx_eq!(f32, atom1.get_velocity().z, atom2.get_velocity().z);
-
-                assert_approx_eq!(f32, atom1.get_force().x, atom2.get_force().x);
-                assert_approx_eq!(f32, atom1.get_force().y, atom2.get_force().y);
-                assert_approx_eq!(f32, atom1.get_force().z, atom2.get_force().z);
+                compare_atoms(atom1, atom2);
             }
         }
     }
@@ -880,23 +925,7 @@ mod tests {
             let frame2 = raw_frame2.unwrap();
 
             for (atom1, atom2) in frame1.atoms_iter().zip(frame2.atoms_iter()) {
-                assert_eq!(atom1.get_residue_number(), atom2.get_residue_number());
-                assert_eq!(atom1.get_residue_name(), atom2.get_residue_name());
-                assert_eq!(atom1.get_atom_number(), atom2.get_atom_number());
-                assert_eq!(atom1.get_atom_name(), atom2.get_atom_name());
-                assert_eq!(atom1.get_chain(), atom2.get_chain());
-
-                assert_approx_eq!(f32, atom1.get_position().x, atom2.get_position().x);
-                assert_approx_eq!(f32, atom1.get_position().y, atom2.get_position().y);
-                assert_approx_eq!(f32, atom1.get_position().z, atom2.get_position().z);
-
-                assert_approx_eq!(f32, atom1.get_velocity().x, atom2.get_velocity().x);
-                assert_approx_eq!(f32, atom1.get_velocity().y, atom2.get_velocity().y);
-                assert_approx_eq!(f32, atom1.get_velocity().z, atom2.get_velocity().z);
-
-                assert_approx_eq!(f32, atom1.get_force().x, atom2.get_force().x);
-                assert_approx_eq!(f32, atom1.get_force().y, atom2.get_force().y);
-                assert_approx_eq!(f32, atom1.get_force().z, atom2.get_force().z);
+                compare_atoms(atom1, atom2);
             }
         }
 
@@ -993,6 +1022,80 @@ mod tests {
             Ok(_) => panic!("Iterator should not have been constructed."),
             Err(ReadTrajError::StartNotFound(_)) => (),
             Err(e) => panic!("Incorrect error type {} returned", e),
+        }
+    }
+
+    #[test]
+    fn read_trr_step_1() {
+        let mut system1 = System::from_file("test_files/example.gro").unwrap();
+        let mut system2 = System::from_file("test_files/example.gro").unwrap();
+
+        for (raw1, raw2) in system1
+            .trr_iter("test_files/short_trajectory.trr")
+            .unwrap()
+            .with_step(1)
+            .unwrap()
+            .zip(system2.trr_iter("test_files/short_trajectory.trr").unwrap())
+        {
+            let frame1 = raw1.unwrap();
+            let frame2 = raw2.unwrap();
+
+            for (atom1, atom2) in frame1.atoms_iter().zip(frame2.atoms_iter()) {
+                compare_atoms(atom1, atom2);
+            }
+        }
+    }
+
+    #[test]
+    fn read_trr_step_3() {
+        let mut system1 = System::from_file("test_files/example.gro").unwrap();
+        let mut system2 = System::from_file("test_files/example.gro").unwrap();
+
+        for (raw1, raw2) in system1
+            .trr_iter("test_files/short_trajectory.trr")
+            .unwrap()
+            .with_step(3)
+            .unwrap()
+            .zip(system2.trr_iter("test_files/short_trajectory.trr").unwrap().step_by(3))
+        {
+            let frame1 = raw1.unwrap();
+            let frame2 = raw2.unwrap();
+
+            for (atom1, atom2) in frame1.atoms_iter().zip(frame2.atoms_iter()) {
+                compare_atoms(atom1, atom2);
+            }
+        }
+    }
+
+    #[test]
+    fn read_trr_step_23() {
+        let mut system1 = System::from_file("test_files/example.gro").unwrap();
+        let mut system2 = System::from_file("test_files/example.gro").unwrap();
+
+        for (raw1, raw2) in system1
+            .trr_iter("test_files/short_trajectory.trr")
+            .unwrap()
+            .with_step(23)
+            .unwrap()
+            .zip(system2.trr_iter("test_files/short_trajectory.trr").unwrap().step_by(23))
+        {
+            let frame1 = raw1.unwrap();
+            let frame2 = raw2.unwrap();
+
+            for (atom1, atom2) in frame1.atoms_iter().zip(frame2.atoms_iter()) {
+                compare_atoms(atom1, atom2);
+            }
+        }
+    }
+
+    #[test]
+    fn read_trr_step_0() {
+        let mut system = System::from_file("test_files/example.gro").unwrap();
+
+        match system.trr_iter("test_files/short_trajectory.trr").unwrap().with_step(0) {
+            Ok(_) => panic!("Should have failed."),
+            Err(ReadTrajError::InvalidStep(s)) => assert_eq!(s, 0),
+            Err(e) => panic!("Incorrect error type {} returned.", e),
         }
     }